+++ client/package.json
@@ -15,6 +15,7 @@
         "apollo-link-http": "^1.5.5",
         "babel-core": "^6.26.3",
         "concurrently": "^4.0.1",
+        "firebase": "^5.9.4",
         "graphql": "^14.0.2",
         "graphql-tag": "^2.10.0",
         "less": "^3.8.1",
@@ -27,6 +28,7 @@
         "react-datepicker": "1.3.0",
         "react-dates": "^18.2.2",
         "react-dom": "^16.6.0",
+        "react-firestore": "^1.0.1",
         "react-ga": "^2.5.3",
         "react-moment": "^0.8.2",
         "react-numeric-input": "^2.2.3",
@@ -50,13 +52,13 @@
     },
     "devDependencies": {
         "@types/graphql": "^14.0.3",
-        "@types/react-dates": "^17.1.3",
         "@types/jest": "^23.3.8",
         "@types/lodash.flowright": "^3.5.4",
         "@types/moment-timezone": "^0.5.9",
         "@types/node": "^10.12.0",
         "@types/prop-types": "^15.5.6",
         "@types/react-datepicker": "^1.1.7",
+        "@types/react-dates": "^17.1.3",
         "@types/react-dom": "^16.0.9",
         "@types/react-numeric-input": "^2.2.3",
         "@types/react-router-dom": "^4.3.1",
