+++ client/src/components/includes/SessionView.tsx
@@ -3,122 +3,25 @@ import * as React from 'react';
 import TopBar from '../includes/TopBar';
 import SessionInformationHeader from '../includes/SessionInformationHeader';
 import SessionQuestionsContainer from '../includes/SessionQuestionsContainer';
-import { Interval } from '../../utilities/interval';
-
-import gql from 'graphql-tag';
-import { Query, Mutation } from 'react-apollo';
 import { Icon } from 'semantic-ui-react';
 
-const GET_SESSION_DATA = gql`
-query getDataForSession($sessionId: Int!, $courseId: Int!) {
-    apiGetCurrentUser {
-        nodes {
-            computedName
-            computedAvatar
-            userId
-            courseUsersByUserId(condition: {courseId: $courseId}) {
-                nodes {
-                    role
-                }
-            }
-            questionsByAskerId {
-                nodes {
-                    status
-                    sessionBySessionId {
-                        sessionId
-                        endTime
-                    }
-                }
-            }
-        }
-    }
-    courseByCourseId(courseId: $courseId) {
-        name
-        code
-        queueOpenInterval {
-            seconds
-            minutes
-            hours
-            days
-            months
-            years
-        }
-    }
-    sessionBySessionId(sessionId: $sessionId) {
-        sessionId
-        startTime
-        endTime
-        building
-        room
-        title
-        questionsBySessionId {
-            nodes {
-                questionId
-                content
-                status
-                timeEntered
-                location
-                userByAskerId {
-                    computedName
-                    computedAvatar
-                    userId
-                }
-                userByAnswererId {
-                    computedName
-                    userId
-                }
-                questionTagsByQuestionId {
-                    nodes {
-                        tagByTagId {
-                            name
-                            level
-                            tagId
-                        }
-                    }
-                }
-            }
-        }
-        sessionTasBySessionId {
-            nodes {
-                userByUserId {
-                    computedName
-                    computedAvatar
-                }
-            }
-        }
-    }
-}
-`;
+import { firestore } from '../includes/firebase';
 
-interface SessionData {
-    sessionBySessionId: AppSession;
-    courseByCourseId: AppCourseInterval;
-    apiGetCurrentUser: {
-        nodes: [AppUserRoleQuestions]
-    };
-}
-
-interface Variables {
-    sessionId: number;
-    courseId: number;
-}
-
-class SessionDataQuery extends Query<SessionData, Variables> { }
-
-const UNDO_QUESTION = gql`
-mutation UndoQuestion($questionId: Int!, $status: String!) {
-    updateQuestionByQuestionId(input: {questionPatch: {status: $status, timeAddressed: null, answererId: null},
-        questionId: $questionId}) {
-        clientMutationId
-    }
-}
-`;
+// const UNDO_QUESTION = gql`
+// mutation UndoQuestion($questionId: Int!, $status: String!) {
+//     updateQuestionByQuestionId(input: {questionPatch: {status: $status, timeAddressed: null, answererId: null},
+//         questionId: $questionId}) {
+//         clientMutationId
+//     }
+// }
+// `;
 
 class SessionView extends React.Component {
     props: {
-        id: number,
-        courseId: number,
+        id: string,
+        course: FireCourse,
         isDesktop: boolean,
+        session: FireSession
         backCallback: Function,
         joinCallback: Function,
     };
@@ -140,6 +43,13 @@ class SessionView extends React.Component {
             undoQuestionId: undefined,
             timeoutId: null,
         };
+
+        firestore
+            .collection('sessions')
+            .doc(this.props.id)
+            .onSnapshot((doc) => {
+                this.setState({ session: doc.data() });
+            });
     }
 
     triggerUndo = (questionId: number, action: string, name: string) => {
@@ -180,17 +90,17 @@ class SessionView extends React.Component {
         });
     }
 
-    isOpen = (session: AppSession, interval: AppInterval): boolean => {
-        return new Date(session.startTime).getTime() - Interval.toMillisecoonds(interval) < new Date().getTime()
+    isOpen = (session: FireSession, interval: number): boolean => {
+        return new Date(session.startTime).getTime() - interval * 1000 < new Date().getTime()
             && new Date(session.endTime) > new Date();
     }
 
-    isPast = (session: AppSession): boolean => {
+    isPast = (session: FireSession): boolean => {
         return new Date() > new Date(session.endTime);
     }
 
-    getOpeningTime = (session: AppSession, interval: AppInterval): Date => {
-        return new Date(new Date(session.startTime).getTime() - Interval.toMillisecoonds(interval));
+    getOpeningTime = (session: FireSession, interval: number): Date => {
+        return new Date(new Date(session.startTime).getTime() - interval * 1000);
     }
 
     render() {
@@ -207,99 +117,90 @@ class SessionView extends React.Component {
             }
         }
 
+        // var otherQuestions = data.apiGetCurrentUser.nodes[0].questionsByAskerId.nodes
+        // // .filter((question) => question.sessionBySessionId.sessionId !== this.props.id)
+        // .filter((question) => question.status === 'unresolved')
+        // .filter((question) => new Date(question.sessionBySessionId.endTime) >= new Date());
+
+
         return (
             <section className="StudentSessionView">
-                <SessionDataQuery
-                    query={GET_SESSION_DATA}
-                    variables={{ sessionId: this.props.id, courseId: this.props.courseId }}
-                    pollInterval={5000}
-                >
-                    {({ loading, data, error, refetch }) => {
-                        if (error) { return null; }
+                {/*
+                if (error) { return null; }
                         if (!data || !data.apiGetCurrentUser) {
                             return null;
-                        }
-                        var otherQuestions = data.apiGetCurrentUser.nodes[0].questionsByAskerId.nodes
-                            .filter((question) => question.sessionBySessionId.sessionId !== this.props.id)
-                            .filter((question) => question.status === 'unresolved')
-                            .filter((question) => new Date(question.sessionBySessionId.endTime) >= new Date());
+                        } */}
 
-                        return (
-                            <React.Fragment>
-                                {this.props.isDesktop &&
-                                    <TopBar
-                                        user={data.apiGetCurrentUser.nodes[0]}
-                                        role={data.apiGetCurrentUser.nodes[0].courseUsersByUserId.nodes[0].role}
-                                        context="session"
-                                        courseId={this.props.courseId}
-                                    />
-                                }
-                                {this.props.id === -1 || !data.sessionBySessionId
-                                    ? <React.Fragment>
-                                        <p className="welcomeMessage">Welcome, <span className="welcomeName">
-                                            {data.apiGetCurrentUser.nodes[0].computedName}
-                                        </span></p>
-                                        <p className="noSessionSelected">
-                                            Please select an office hour from the calendar.
+                {this.props.isDesktop &&
+                    <TopBar
+                        user={data.apiGetCurrentUser.nodes[0]}
+                        role={data.apiGetCurrentUser.nodes[0].courseUsersByUserId.nodes[0].role}
+                        context="session"
+                        courseId={this.props.courseId}
+                    />
+                }
+                {this.props.id === '-1' || !data.sessionBySessionId
+                    ? <React.Fragment>
+                        <p className="welcomeMessage">Welcome, <span className="welcomeName">
+                            {data.apiGetCurrentUser.nodes[0].computedName}
+                        </span></p>
+                        <p className="noSessionSelected">
+                            Please select an office hour from the calendar.
                                         </p>
-                                    </React.Fragment>
-                                    : <React.Fragment>
-                                        <SessionInformationHeader
-                                            session={data.sessionBySessionId}
-                                            course={data.courseByCourseId}
-                                            myUserId={data.apiGetCurrentUser.nodes[0].userId}
-                                            callback={this.props.backCallback}
-                                            isDesktop={this.props.isDesktop}
-                                        />
-                                        {this.state.undoQuestionId &&
-                                            <Mutation mutation={UNDO_QUESTION} onCompleted={this.dismissUndo}>
-                                                {(undoQuestion) =>
-                                                    <div className="undoContainer">
-                                                        <p className="undoClose" onClick={this.dismissUndo}>
-                                                            <Icon name="close" />
-                                                        </p>
-                                                        <p className="undoText">
-                                                            {undoText}
-                                                            <span
-                                                                className="undoLink"
-                                                                onClick={() =>
-                                                                    this.handleUndoClick(undoQuestion, refetch)
-                                                                }
-                                                            >
-                                                                Undo
-                                                            </span>
-                                                        </p>
-                                                    </div>
+                    </React.Fragment>
+                    : <React.Fragment>
+                        <SessionInformationHeader
+                            session={data.sessionBySessionId}
+                            course={data.courseByCourseId}
+                            myUserId={data.apiGetCurrentUser.nodes[0].userId}
+                            callback={this.props.backCallback}
+                            isDesktop={this.props.isDesktop}
+                        />
+                        {this.state.undoQuestionId &&
+                            <Mutation mutation={UNDO_QUESTION} onCompleted={this.dismissUndo}>
+                                {(undoQuestion) =>
+                                    <div className="undoContainer">
+                                        <p className="undoClose" onClick={this.dismissUndo}>
+                                            <Icon name="close" />
+                                        </p>
+                                        <p className="undoText">
+                                            {undoText}
+                                            <span
+                                                className="undoLink"
+                                                onClick={() =>
+                                                    this.handleUndoClick(undoQuestion, refetch)
                                                 }
-                                            </Mutation>
-                                        }
-                                        <SessionQuestionsContainer
-                                            isTA={data.apiGetCurrentUser.nodes[0].
-                                                courseUsersByUserId.nodes[0].role !== 'student'}
-                                            questions={data.sessionBySessionId.questionsBySessionId
-                                                .nodes.filter(
-                                                    q => q.status === 'unresolved' || q.status === 'assigned')}
-                                            handleJoinClick={this.props.joinCallback}
-                                            myUserId={data.apiGetCurrentUser.nodes[0].userId}
-                                            triggerUndo={this.triggerUndo}
-                                            refetch={refetch}
-                                            // this sets a ref, which allows a parent to call methods on a child.
-                                            // Here, the parent can't access refetch, but the child can.
-                                            ref={(ref) => this.questionsContainer = ref}
-                                            isOpen={this.isOpen(
-                                                data.sessionBySessionId,
-                                                data.courseByCourseId.queueOpenInterval)}
-                                            isPast={this.isPast(data.sessionBySessionId)}
-                                            openingTime={this.getOpeningTime(
-                                                data.sessionBySessionId, data.courseByCourseId.queueOpenInterval)}
-                                            haveAnotherQuestion={otherQuestions.length > 0}
-                                        />
-                                    </React.Fragment>
+                                            >
+                                                Undo
+                                            </span>
+                                        </p>
+                                    </div>
                                 }
-                            </React.Fragment>
-                        );
-                    }}
-                </SessionDataQuery>
+                            </Mutation>
+                        }
+                        <SessionQuestionsContainer
+                            isTA={data.apiGetCurrentUser.nodes[0].
+                                courseUsersByUserId.nodes[0].role !== 'student'}
+                            questions={data.sessionBySessionId.questionsBySessionId
+                                .nodes.filter(
+                                    q => q.status === 'unresolved' || q.status === 'assigned')}
+                            handleJoinClick={this.props.joinCallback}
+                            myUserId={data.apiGetCurrentUser.nodes[0].userId}
+                            triggerUndo={this.triggerUndo}
+                            refetch={refetch}
+                            // this sets a ref, which allows a parent to call methods on a child.
+                            // Here, the parent can't access refetch, but the child can.
+                            ref={(ref) => this.questionsContainer = ref}
+                            isOpen={this.isOpen(
+                                data.sessionBySessionId,
+                                data.courseByCourseId.queueOpenInterval)}
+                            isPast={this.isPast(data.sessionBySessionId)}
+                            openingTime={this.getOpeningTime(
+                                data.sessionBySessionId, data.courseByCourseId.queueOpenInterval)}
+                            haveAnotherQuestion={otherQuestions.length > 0}
+                        />
+                    </React.Fragment>
+                }
             </section>
         );
     }
