+++ client/src/components/pages/ProfessorDashboardView.tsx
@@ -33,7 +33,7 @@ interface ProfessorMetadataData {
 }
 
 interface MetadataVariables {
-    courseId: number;
+    courseId: string;
 }
 
 class ProfessorMetadataDataQuery extends Query<ProfessorMetadataData, MetadataVariables> { }
@@ -53,14 +53,12 @@ class ProfessorDashboardView extends React.Component {
     }
 
     render() {
-        let courseId = parseInt(this.props.match.params.courseId, 10);
+        let courseId = this.props.match.params.courseId;
         return (
             <div className="ProfessorView">
                 <ProfessorMetadataDataQuery
                     query={METADATA_QUERY}
-                    variables={{
-                        courseId: courseId
-                    }}
+                    variables={{ courseId: '2' }}
                 >
                     {({ loading, data }) => {
                         var courseCode: string = 'Loading...';
